% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{checker}
\alias{checker}
\title{Check User Answer}
\usage{
checker(
  label,
  user_code,
  solution_code,
  check_code,
  envir_result,
  evaluate_result,
  envir_prep,
  ...
)
}
\arguments{
\item{label}{}

\item{user_code}{}

\item{solution_code}{}

\item{check_code}{}

\item{envir_result}{}

\item{evaluate_result}{}

\item{envir_prep}{}

\item{...}{}
}
\value{

}
\description{
Check User Answer
}
