% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_learnr.R
\name{get_tutorial_path}
\alias{get_tutorial_path}
\title{Copied from learnr}
\usage{
get_tutorial_path(name, package)
}
\arguments{
\item{name}{}

\item{package}{}
}
\value{

}
\description{
Copied from learnr
}
