% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_learnr.R
\name{install_tutorial_dependencies}
\alias{install_tutorial_dependencies}
\title{Copied from learnr}
\usage{
install_tutorial_dependencies(dir)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Copied from learnr
}
