% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\title{Read Chapter Notes}
\usage{
notes(chapter)
}
\arguments{
\item{chapter}{The number of the chapter to read.}
}
\value{
A window will display with the notes for the selected chapter.
}
\description{
Read Chapter Notes
}
