% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_learnr.R
\name{run_tut}
\alias{run_tut}
\title{Custom run_tutorial function from learnr to accept and pass render_args.}
\usage{
run_tut(name = NULL, package = NULL, shiny_args = NULL, render_args = NULL)
}
\arguments{
\item{name}{}

\item{package}{}

\item{shiny_args}{}

\item{render_args}{}
}
\value{

}
\description{
Custom run_tutorial function from learnr to accept and pass render_args.
}
